refactor platform errors into controller
