+++ crates/controller/src/joystick/error.rs
@@ -3,6 +3,7 @@ use crate::{AxisError, ButtonError, PolarityError, StateError};
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub enum JoyStickError {
   Axis(AxisError),
+  UnsupportedCode(u16),
   InvalidState(StateError),
   InvalidPolarity(PolarityError),
   Button(ButtonError),
@@ -17,6 +18,12 @@ impl From<PolarityError> for JoyStickError {
 impl From<AxisError> for JoyStickError {
   fn from(value: AxisError) -> Self {
     Self::Axis(value)
+  }
+}
+
+impl From<u16> for JoyStickError {
+  fn from(v: u16) -> Self {
+    Self::UnsupportedCode(v)
   }
 }
 
@@ -37,6 +44,9 @@ impl std::fmt::Display for JoyStickError {
     match self {
       JoyStickError::Axis(axis_error) => {
         writeln!(f, "{}", axis_error)
+      }
+      JoyStickError::UnsupportedCode(code) => {
+        writeln!(f, "Joystick unsupported code: {:#?}", code)
       }
       JoyStickError::Button(button_error) => {
         writeln!(f, "{}", button_error)
